re-refactor as per liviu's suggestions to split up retreive and transformation logic
