+++ bia-export/bia_export/website_export/image_pages/retrieve.py
@@ -1,44 +1,16 @@
-from bia_export.website_export.utils import read_api_json_file
-from bia_export.website_export.utils import (
-    read_all_json,
-)
-from bia_export.website_export.website_models import (
-    BioSample,
-    ExperimentallyCapturedImage,
-    ImageAcquisition,
-    Specimen,
-    SpecimenGrowthProtocol,
-    SpecimenImagingPreparationProtocol,
-    StudyCreationContext,
-)
 from uuid import UUID
-from bia_shared_datamodels import bia_data_model
 from pydantic import BaseModel
+from bia_export.website_export.utils import read_all_json, read_api_json_file
+from bia_export.website_export.website_models import CLIContext
+from bia_shared_datamodels import bia_data_model
 from typing import List, Type
 import logging
 
 logger = logging.getLogger("__main__." + __name__)
 
 
-def create_ec_images(context: StudyCreationContext) -> ExperimentallyCapturedImage:
-    eci_map = {}
-
-    api_images = retrieve_images(context)
-
-    if context.root_directory:
-        context.image_to_canonical_rep_uuid_map = get_local_img_rep_map(context)
-
-    for api_image in api_images:
-
-        image = create_image(api_image, context)
-
-        eci_map[str(image.uuid)] = image.model_dump(mode="json")
-
-    return eci_map
-
-
 def retrieve_images(
-    context: StudyCreationContext,
+    context: CLIContext,
 ) -> list[bia_data_model.ImageRepresentation]:
 
     if context.root_directory:
@@ -56,57 +28,8 @@ def retrieve_images(
     return api_ecis
 
 
-def create_image(
-    api_image: bia_data_model.ExperimentallyCapturedImage, context: StudyCreationContext
-) -> ExperimentallyCapturedImage:
-    website_fields = {}
-    api_image_acquisitions = retrieve_object_list(
-        api_image.acquisition_process_uuid, bia_data_model.ImageAcquisition, context
-    )
-
-    website_fields["acquisition_process"] = create_details_list(
-        api_image_acquisitions, ImageAcquisition
-    )
-
-    api_specimen = retrieve_specimen(api_image.subject_uuid, context)
-    api_biosamples = retrieve_object_list(
-        api_specimen.sample_of_uuid, bia_data_model.BioSample, context
-    )
-    api_specimen_growth_protocols = retrieve_object_list(
-        api_specimen.growth_protocol_uuid,
-        bia_data_model.SpecimenGrowthProtocol,
-        context,
-    )
-    api_specimen_imaging_preparation_protocols = retrieve_object_list(
-        api_specimen.imaging_preparation_protocol_uuid,
-        bia_data_model.SpecimenImagingPreparationProtocol,
-        context,
-    )
-
-    specimen_dict = api_specimen.model_dump() | {
-        "imaging_preparation_protocol": create_details_list(
-            api_specimen_imaging_preparation_protocols,
-            SpecimenImagingPreparationProtocol,
-        ),
-        "sample_of": create_details_list(api_biosamples, BioSample),
-        "growth_protocol": create_details_list(
-            api_specimen_growth_protocols, SpecimenGrowthProtocol
-        ),
-    }
-    website_fields["subject"] = Specimen(**specimen_dict)
-
-    canonical_img_rep = retrieve_canonical_representation(api_image.uuid, context)
-
-    image_dict = (
-        api_image.model_dump()
-        | website_fields
-        | {"canonical_representation": canonical_img_rep}
-    )
-    return ExperimentallyCapturedImage(**image_dict)
-
-
 def retrieve_specimen(
-    specimen_uuid: UUID, context: StudyCreationContext
+    specimen_uuid: UUID, context: CLIContext
 ) -> bia_data_model.Specimen:
     if context.root_directory:
         specimen_path = context.root_directory.joinpath(
@@ -122,7 +45,7 @@ def retrieve_specimen(
 
 
 def retrieve_object_list(
-    uuid_list: list[UUID], api_class: Type[BaseModel], context: StudyCreationContext
+    uuid_list: list[UUID], api_class: Type[BaseModel], context: CLIContext
 ) -> List[BaseModel]:
     if context.root_directory:
 
@@ -147,18 +70,8 @@ def retrieve_object_list(
     return obj_list
 
 
-def create_details_list(
-    api_object_list: List[BaseModel], website_class: Type[BaseModel]
-):
-    obj_list = []
-    for api_obj in api_object_list:
-        obj_dict = api_obj.model_dump() | {"default_open": True}
-        obj_list.append(website_class(**obj_dict))
-    return obj_list
-
-
 def retrieve_canonical_representation(
-    image_uuid: UUID, context: StudyCreationContext
+    image_uuid: UUID, context: CLIContext
 ) -> bia_data_model.ImageRepresentation:
     if context.root_directory:
         canonical_rep = context.root_directory.joinpath(
@@ -182,7 +95,7 @@ def is_canonical(image_rep: bia_data_model.ImageRepresentation) -> bool:
         return False
 
 
-def get_local_img_rep_map(context: StudyCreationContext) -> dict[UUID, UUID]:
+def get_local_img_rep_map(context: CLIContext) -> dict[UUID, UUID]:
     image_rep_path = context.root_directory.joinpath(
         f"image_representations/{context.accession_id}/*.json"
     )
